+++ packages/core-components/src/lib/unsorted/viz/BarChart.svelte
@@ -43,7 +43,15 @@
 		)?.evidenceType;
 	});
 
-	export let showAllXAxisLabels = xType === 'category' || xEvidenceType === 'string';
+	$: if (!showAllXaxisLabelsManuallySet)
+		showAllXAxisLabels = xType === 'category' || xEvidenceType === 'string';
+
+	/** @type {boolean} */
+	export let showAllXAxisLabels;
+	const showAllXaxisLabelsManuallySet = typeof showAllXAxisLabels !== 'undefined';
+
+	$: if (typeof showAllXAxisLabels === 'string')
+		showAllXAxisLabels = showAllXAxisLabels?.toLowerCase() === 'true';
 
 	$: {
 		if (swapXY === 'true' || swapXY === true) {