+++ packages/core-components/src/lib/unsorted/viz/ErrorChart.svelte
@@ -12,7 +12,7 @@
 
 <div
 	width="100%"
-	class="grid grid-rows-auto box-content grid-cols-1 justify-center bg-red-50 text-grey-700 font-ui font-normal rounded border border-red-100 min-h-[150px] py-5 px-8 my-5 print:break-inside-avoid"
+	class="grid grid-rows-auto box-content grid-cols-1 justify-center bg-red-50 text-grey-700 font-ui font-normal rounded border border-red-200 min-h-[150px] py-5 px-8 my-5 print:break-inside-avoid"
 >
 	<div class="m-auto w-full">
 		<div class="font-bold text-center text-lg">{chartType}</div>